add an exception field for tags, so they will be stripped away
this was needed, because with the new option to leave text in place instead of a label on todoist, sometimes you want to remove more tags than only #todoist, e.g. for eisenhower-matrix. So this is the way to remove this tags... Or you want to see the text todoist on all your tasks. Maybe useful.